requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.7
django-debug-toolbar==6.1.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2==2.9.11
PyYAML==6.0.3
referencing==0.37.0
rest-framework-simplejwt==0.0.2
rpds-py==0.30.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
